api/requirements.txt
Flask
flask_cors
numpy
scipy
pandas
jams
librosa
keras
tensorflow-cpu
musicxml
gunicorn
python-dotenv
azure-storage-blob
pyodbc
pyjwt